Switch line borders from bottom to top
Previously borders on lines were on the bottom of the current
one, and this change switches it to the top of the next one.
This is useful so that where a junction between a code line
and a blank line happens, we can use the + operator (next
immediate sibling) to turn on a missing line.

We can't do this the other way around as CSS has no previous
immediate sibling operator.

I've left the new line in as red for debugging purposes,
will remove that in another commit.